# ENH: Improve performance of DE's `_ensure_constraint`