--- a/scipy/optimize/_differentialevolution.py
+++ b/scipy/optimize/_differentialevolution.py
@@ -689,9 +689,8 @@ class DifferentialEvolutionSolver(object):
         """
         make sure the parameters lie between the limits
         """
-        for index, param in enumerate(trial):
-            if param > 1 or param < 0:
-                trial[index] = self.random_number_generator.rand()
+        values = self.random_number_generator.rand(trial.size)
+        np.putmask(trial, (trial < 0) | (trial > 1), values)
 
     def _mutate(self, candidate):
         """