@andyfaff 2017-11-13
@mscavnicky thank you for contributing to scipy for the first time.
The test fail on travis is unrelated to this PR. 

Most of the problems I use `differential_evolution` for probably only have on the order of 10 parameters. At that level the difference between the two approaches you mention is negligible.
(my objective function normally takes an order of magnitude longer than any overhead incurred in scipy). Still, it's good to improve performance.

Unfortunately the improvement you suggest has a few issues with it. The main one is that for all the entries that are outside the range, it uses the same random number as the replacement.

```
import numpy as np
def f(trial):
    res = np.copy(trial)
    for index, param in enumerate(trial):
        if param > 1 or param < 0:
            res[index] = np.random.rand()
    return res

def f2(trial):
    res = np.copy(trial)
    for index in np.where((trial < 0) | (trial > 1)):
        res[index] = np.random.rand()
    return res

ll = -1
ul = 2
trial = (ul - ll) * np.random.rand((10)) + ll
print('trial', trial)
print('original', f(trial))
print('suggested', f2(trial)
```
gives the following:
```
trial [ 1.54373367  0.59144653 -0.12993146 -0.99878989 -0.61958958  0.47539014
  0.81057796  0.15478569 -0.02327803  1.16833128]
original [ 0.09154946  0.59144653  0.57734997  0.08478023  0.19129296  0.47539014
  0.81057796  0.15478569  0.5429328   0.70491116]
suggested [ 0.01449935  0.59144653  0.01449935  0.01449935  0.01449935  0.47539014
  0.81057796  0.15478569  0.01449935  0.01449935]
```
You can see in the last one that the replacement entries are all the same.

A better replacement for the code is:
```
def f3(trial):
      return np.where((trial < 0) | (trial > 1), np.random.rand(trial.size), trial)
```

which makes the scipy replacement:
```
trial = np.where((trial < 0) | (trial > 1), np.random.rand(trial.size), trial)
```

Thus, we can get rid of the for loop and the if test. I wonder how this changes the timings?