@mscavnicky 2017-11-13
Thanks for catching the error. That was a basic mistake on my part.

The method that you have proposed seems to be slightly slower especially in the larger dimensions since it needs to allocate and generate the whole random array of the size of the trial.

```trial = np.where((trial < 0) | (trial > 1), np.random.rand(trial.size), trial)```

By replacing `np.where` with `np.putmask` we can get rid of the assignment at least. That is a micro-optimisation but seems to consistently perform better.

```np.putmask(trial, (trial < 0) | (trial > 1), np.random.rand(trial.size))```

The new timings are in the table below. The structure of the table is changed to make it more readable (every method now has its own column). I have also added extra rows for arrays with 2 elements.

```
dim	prob	scipy	np.where np.putmask
2	0.001	1.89	5.75	 4.51
2	0.005	2.15	5.98	 4.59
2	0.01	1.94	6.02	 5.03
10	0.001	5.23	5.83	 4.9
10	0.005	5.15	6.11	 4.82
10	0.01	5.22	5.89	 4.98
100	0.001	42.4	7.3	 6.13
100	0.005	42.5	7.03	 6.08
100	0.01	44.8	7.26	 6.04
1000	0.001	427	20	 19.3
1000	0.005	428	20.2	 18.8
1000	0.01	427	19.8	 18.7
```

The current implementation performs better in low dimensions which you mention are your general use case, but its complexity grows linearly. The method using `np.putmask` has some fixed costs but has a steady performance with growing dimension. My current problem has 50-100 dimensions and the ~5x speedup is helpful, given that this method is called for every instance.

Ultimately, this is a trade-off between keeping DE optimized for smaller dimensions or making it more robust towards n-dimensional problems and introducing some fixed cost.