@mscavnicky 2017-11-13
Actually, my original reimplementation was incorrectly using the result of `np.where` which returns a tuple. It can be fixed and gives better performance than the methods from my previous comment. See below:

```
def fixed(trial):
  for index in np.where((trial < 0) | (trial > 1))[0]:
    trial[index] = self.random_number_generator.rand()
```

```
dim	prob	scipy	np.where	np.putmask	fixed
2	0.001	1.96	6		4.66		4.39
2	0.005	1.93	6.07		4.76		4.36
2	0.01	1.98	5.84		4.69		4.39
10	0.001	5.29	6.25		4.83		4.39
10	0.005	5.18	6.46		4.71		4.57
10	0.01	5.2	6.36		4.82		4.58
100	0.001	42.8	7.23		5.97		4.44
100	0.005	45.8	7.23		6.03		4.46
100	0.01	42.9	7.31		5.86		4.41
1000	0.001	431	20.2		19.4		5.37
1000	0.005	430	20.5		18.6		5.51
1000	0.01	428	20.1		18.8		5.59
```

The conclusion of the previous comment still holds.