@andyfaff 2017-11-13
So it seems like the last suggestion is the way to go (although I like the one-liner). Could you change the PR to that please?