@andyfaff 2017-11-20
Optimisation applied in #8177